wip on frozen contract decoding
