+++ zkvm/src/lib.rs
@@ -1,16 +1,7 @@
-extern crate byteorder;
-extern crate core;
-extern crate rand;
-
-extern crate bulletproofs;
-extern crate curve25519_dalek;
-extern crate merlin;
-extern crate spacesuit;
-extern crate subtle;
-
 #[macro_use]
 extern crate failure;
 
+mod contract;
 mod encoding;
 mod errors;
 mod ops;
@@ -24,8 +15,12 @@ mod types;
 mod verifier;
 mod vm;
 
+pub use self::contract::{Contract, FrozenContract, FrozenItem, FrozenValue, Input, PortableItem};
 pub use self::errors::VMError;
+pub use self::ops::{Instruction, Opcode};
+pub use self::predicate::{Predicate, PredicateWitness};
 pub use self::prover::Prover;
 pub use self::txlog::{Entry, TxID, UTXO};
+pub use self::types::{Commitment, Data, Item, Value, WideValue};
 pub use self::verifier::Verifier;
 pub use self::vm::{Tx, VerifiedTx};
