+++ zkvm/src/transcript.rs
@@ -1,6 +1,5 @@
 //! Defines a `TranscriptProtocol` trait for using a Merlin transcript.
 
-use byteorder::{ByteOrder, LittleEndian};
 use curve25519_dalek::ristretto::CompressedRistretto;
 use curve25519_dalek::scalar::Scalar;
 use merlin::Transcript;
@@ -12,12 +11,6 @@ pub trait TranscriptProtocol {
     fn commit_point(&mut self, label: &'static [u8], point: &CompressedRistretto);
     /// Compute a `label`ed challenge variable.
     fn challenge_scalar(&mut self, label: &'static [u8]) -> Scalar;
-}
-
-fn le_u64(value: u64) -> [u8; 8] {
-    let mut value_bytes = [0u8; 8];
-    LittleEndian::write_u64(&mut value_bytes, value);
-    value_bytes
 }
 
 impl TranscriptProtocol for Transcript {
