+++ zkvm/src/types.rs
@@ -6,13 +6,12 @@ use curve25519_dalek::scalar::Scalar;
 use merlin::Transcript;
 use spacesuit::SignedInteger;
 
-use crate::encoding;
+use crate::contract::{Contract, Input, InputWitness, PortableItem};
 use crate::encoding::Subslice;
 use crate::errors::VMError;
 use crate::ops::Instruction;
 use crate::predicate::Predicate;
 use crate::transcript::TranscriptProtocol;
-use crate::txlog::{TxID, UTXO};
 
 #[derive(Debug)]
 pub enum Item {
@@ -25,22 +24,10 @@ pub enum Item {
     Constraint(Constraint),
 }
 
-#[derive(Debug)]
-pub enum PortableItem {
-    Data(Data),
-    Value(Value),
-}
-
 #[derive(Clone, Debug)]
 pub enum Data {
     Opaque(Vec<u8>),
     Witness(DataWitness),
-}
-
-#[derive(Debug)]
-pub struct Contract {
-    pub(crate) payload: Vec<PortableItem>,
-    pub(crate) predicate: Predicate,
 }
 
 #[derive(Debug)]
@@ -84,19 +71,6 @@ pub enum Commitment {
     Open(Box<CommitmentWitness>),
 }
 
-#[derive(Clone, Debug)]
-pub enum Input {
-    Opaque(Vec<u8>),
-    Witness(Box<InputWitness>),
-}
-
-#[derive(Clone, Debug)]
-pub struct InputWitness {
-    contract: FrozenContract,
-    utxo: UTXO,
-    txid: TxID,
-}
-
 /// Prover's representation of the witness.
 #[derive(Clone, Debug)]
 pub enum DataWitness {
@@ -112,29 +86,6 @@ pub enum DataWitness {
 pub struct CommitmentWitness {
     pub value: ScalarWitness,
     pub blinding: Scalar,
-}
-
-/// Representation of a Contract inside an Input that can be cloned.
-#[derive(Clone, Debug)]
-pub struct FrozenContract {
-    pub(crate) payload: Vec<FrozenItem>,
-    pub(crate) predicate: Predicate,
-}
-
-/// Representation of a PortableItem inside an Input that can be cloned.
-#[derive(Clone, Debug)]
-pub enum FrozenItem {
-    Data(Data),
-    Value(FrozenValue),
-}
-
-/// Representation of a Value inside an Input that can be cloned.
-/// Note: values do not necessarily have open commitments. Some can be reblinded,
-/// others can be passed-through to an output without going through `cloak` and the constraint system.
-#[derive(Clone, Debug)]
-pub struct FrozenValue {
-    pub(crate) qty: Commitment,
-    pub(crate) flv: Commitment,
 }
 
 /// Represents a concrete kind of a number represented by a scalar:
@@ -160,10 +111,6 @@ impl Commitment {
             Commitment::Closed(x) => Ok(*x),
         }
     }
-
-    fn encode(&self, buf: &mut Vec<u8>) {
-        buf.extend_from_slice(&self.to_point().to_bytes());
-    }
 }
 
 impl CommitmentWitness {
@@ -172,15 +119,8 @@ impl CommitmentWitness {
         gens.commit(self.value.into(), self.blinding).compress()
     }
 
-    fn encode(&self, buf: &mut Vec<u8>) {
+    pub(crate) fn encode(&self, buf: &mut Vec<u8>) {
         buf.extend_from_slice(&self.to_point().to_bytes());
-    }
-}
-
-impl InputWitness {
-    fn encode(&self, buf: &mut Vec<u8>) {
-        buf.extend_from_slice(&self.txid.0);
-        self.contract.encode(buf);
     }
 }
 
@@ -350,42 +290,6 @@ impl DataWitness {
             DataWitness::Commitment(cw) => cw.encode(buf),
             DataWitness::Scalar(s) => buf.extend_from_slice(&s.to_bytes()),
             DataWitness::Input(b) => b.encode(buf),
-        }
-    }
-}
-
-impl Contract {
-    pub fn min_serialized_length(&self) -> usize {
-        let mut size = 32 + 4;
-        for item in self.payload.iter() {
-            match item {
-                PortableItem::Data(d) => size += 1 + 4 + d.min_serialized_length(),
-                PortableItem::Value(_) => size += 1 + 64,
-            }
-        }
-        size
-    }
-}
-
-impl FrozenContract {
-    fn encode(&self, buf: &mut Vec<u8>) {
-        buf.extend_from_slice(&self.predicate.point().to_bytes());
-        for p in self.payload.iter() {
-            match p {
-                // Data = 0x00 || LE32(len) || <bytes>
-                FrozenItem::Data(d) => {
-                    buf.push(0u8);
-                    let mut bytes = d.to_bytes();
-                    encoding::write_u32(bytes.len() as u32, buf);
-                    buf.extend_from_slice(&mut bytes);
-                }
-                // Value = 0x01 || <32 bytes> || <32 bytes>
-                FrozenItem::Value(v) => {
-                    buf.push(1u8);
-                    buf.extend_from_slice(&v.qty.to_point().to_bytes());
-                    buf.extend_from_slice(&v.flv.to_point().to_bytes());
-                }
-            }
         }
     }
 }
